pyproject.toml
[project]
name = "technoshare-commentator"
version = "0.1.0"
description = "Local TechnoShare Commentator Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "beautifulsoup4",
    "trafilatura",
    "tenacity",
    "openai",
    "rich",
    "python-dotenv",
    "slack-sdk",
    "pyyaml>=6.0.3",
    "ngrok>=1.7.0",
    "markupsafe>=3.0.3",
]

[project.scripts]
technoshare-ingest = "technoshare_commentator.main_ingest:app"
technoshare-worker = "technoshare_commentator.main_worker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/technoshare_commentator"]
